Refactored a few sections
Created a new class for setting the constraints for the GridBagLayout while setting up the panels. Updated a few logic statements to clean up and make it look better. The 2 biggest changes were the new Constraints.java class and updating my ButtonHandler class to use a switch statement instead of a bunch of if, else if statements.